# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/stadiums": {
            "get": {
                "description": "list stadium entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List stadium entities",
                "operationId": "list-stadium",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Stadium"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create stadium",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create stadium",
                "operationId": "create-stadium",
                "parameters": [
                    {
                        "description": "Stadium entity",
                        "name": "stadium",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Stadium"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Stadium"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/stadiums/{id}": {
            "get": {
                "description": "get stadium by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a stadium entity by ID",
                "operationId": "get-stadium",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stadium ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Stadium"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update stadium by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a stadium entity by ID",
                "operationId": "update-stadium",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stadium ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stadium entity",
                        "name": "stadium",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Stadium"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Stadium"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get stadium by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a stadium entity by ID",
                "operationId": "delete-stadium",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stadium ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Borrow": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BorrowQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BorrowEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "type": {
                    "description": "Type holds the value of the \"type\" field.",
                    "type": "string"
                }
            }
        },
        "ent.BorrowEdges": {
            "type": "object",
            "properties": {
                "borrowConfirmation": {
                    "description": "BorrowConfirmation holds the value of the BorrowConfirmation edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Confirmation"
                    }
                }
            }
        },
        "ent.Confirmation": {
            "type": "object",
            "properties": {
                "adddate": {
                    "description": "Adddate holds the value of the \"adddate\" field.",
                    "type": "string"
                },
                "bookingend": {
                    "description": "Bookingend holds the value of the \"bookingend\" field.",
                    "type": "string"
                },
                "bookingstart": {
                    "description": "Bookingstart holds the value of the \"bookingstart\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ConfirmationQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ConfirmationEdges"
                },
                "hourstime": {
                    "description": "Hourstime holds the value of the \"hourstime\" field.",
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ConfirmationEdges": {
            "type": "object",
            "properties": {
                "confirmationBorrow": {
                    "description": "ConfirmationBorrow holds the value of the ConfirmationBorrow edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Borrow"
                },
                "confirmationStadium": {
                    "description": "ConfirmationStadium holds the value of the ConfirmationStadium edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Stadium"
                },
                "confirmationUser": {
                    "description": "ConfirmationUser holds the value of the ConfirmationUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Stadium": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the StadiumQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.StadiumEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "namestadium": {
                    "description": "Namestadium holds the value of the \"namestadium\" field.",
                    "type": "string"
                }
            }
        },
        "ent.StadiumEdges": {
            "type": "object",
            "properties": {
                "stadiumConfirmation": {
                    "description": "StadiumConfirmation holds the value of the StadiumConfirmation edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Confirmation"
                    }
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "birthday": {
                    "description": "Birthday holds the value of the \"birthday\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                },
                "tel": {
                    "description": "Tel holds the value of the \"tel\" field.",
                    "type": "string"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "userConfirmation": {
                    "description": "UserConfirmation holds the value of the UserConfirmation edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Confirmation"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}